requirements.txt
pytest
azure-common==1.1.27
azure-identity==1.13.0
azure-kusto-data==1.0.3
azure-storage-blob==12.6.0
smart_open==5.2.0
smart_open[azure]==5.2.0
setuptools